Migrate back button to current item ui
