+++ js/components/HomeScene/Nav/Nav.jsx
@@ -69,54 +69,9 @@ const NavArrows = ({ visible }) => {
     );
 };
 
-const NavBack = ({ visible }) => {
-    const el = useRef(null);
-    const resetCurrentBoxUuid = useStore((state) => state.resetCurrentBoxUuid);
-    useLayoutEffect(() => {
-        const target = el.current;
-        if (!target) {
-            return;
-        }
-        gsap.fromTo(
-            target,
-            {
-                opacity: visible ? 0 : 1,
-                y: visible ? 50 : 0,
-                zIndex: visible ? -1 : 1,
-                visibility: visible ? "hidden" : "visible",
-            },
-            {
-                opacity: visible ? 1 : 0,
-                y: visible ? 0 : -50,
-                zIndex: visible ? 1 : -1,
-                visibility: visible ? "visible" : "hidden",
-                duration: 0.2,
-            },
-        );
-        return () => {
-            gsap.killTweensOf(target);
-        };
-    }, [visible]);
-    return (
-        <div className={styles.backWrapper} ref={el}>
-            <NavButton
-                direction="left"
-                onClick={() => {
-                    resetCurrentBoxUuid();
-                }}
-                hotkey="esc"
-            />
-        </div>
-    );
-};
-
 const NavUi = () => {
     const interactable = useStore((state) => state.interactable);
     const currentBoxUuid = useStore((state) => state.currentBoxUuid);
-
-    const isNavBackVisible = useMemo(() => {
-        return typeof currentBoxUuid === "string";
-    }, [currentBoxUuid]);
 
     const isNavArrowsVisible = useMemo(() => {
         return currentBoxUuid === null;
@@ -128,7 +83,6 @@ const NavUi = () => {
 
     return (
         <>
-            <NavBack visible={isNavBackVisible} />
             <NavArrows visible={isNavArrowsVisible} />
         </>
     );
